Fix exception when unpickling operator objects
Discovered when debugging a deadlock when evaluating a roll in a process pool
executor. The use of the negate (-) operator was enough to trigger the bug.

Pickle is used to de/serialize data when communicating with worker processes,
but it was unable to reconstruct the operator object because of missing
arguments to `__new__`. Overriding the `__getnewargs__` method to return
the necessary arguments solved the problem, since the default is assumed to
be an empty tuple.

reprex:
```py
r = dice.roll('-d20', raw=True)
d = pickle.dumps(r)
pickle.loads(d)
```